requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
benepar==0.2.0
blis==0.7.9
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
confection==0.0.4
contourpy==1.0.5
cvxopt==1.3.0
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.3
decorator==5.1.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
entrypoints==0.4
executing==1.1.1
filelock==3.12.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
h5py==3.8.0
huggingface-hub==0.14.1
idna==3.4
imageio==2.28.1
ipykernel==6.16.1
ipympl==0.9.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.3
jupyter_core==4.11.2
jupyterlab-widgets==3.0.3
k3d==2.15.2
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.2
lightning-utilities==0.4.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
msgpack==1.0.5
multidict==6.0.3
murmurhash==1.0.9
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
nptyping==2.5.0
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
parso==0.8.3
path==16.6.0
pathy==0.10.1
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.5.0
pooch==1.7.0
preshed==3.0.8
prompt-toolkit==3.0.31
protobuf==3.20.1
psutil==5.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
Pygments==2.13.0
pynrrd==1.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytorch-lightning==1.8.5
pytz==2022.6
pyvista==0.39.0
PyWavelets==1.4.1
pywin32==304
PyYAML==6.0
pyzmq==24.0.1
regex==2023.5.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.1.3
scipy==1.9.3
scooby==0.7.2
seaborn==0.12.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sklearn==0.0
smart-open==6.3.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-sentence-bert==0.1.2
srsly==2.4.6
stack-data==0.5.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
thinc==8.1.10
threadpoolctl==3.1.0
tifffile==2023.4.12
tokenizers==0.13.3
torch==1.13.0
torch-struct==0.5
torchmetrics==0.11.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
traittypes==0.2.1
transformers==4.28.1
trimesh==3.21.5
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.13
vtk==9.2.6
wasabi==1.1.1
wcwidth==0.2.5
Werkzeug==2.2.2
widgetsnbextension==4.0.3
yarl==1.8.2
